合作机会优化

+++ app/Http/Controllers/Admin/Ks/OpportunityCategoryController.php
@@ -113,6 +113,12 @@ class OpportunityCategoryController extends BaseController
     public function destroy($id)
     {
         //
+        $count=DB::table('cooperation_opportunity_cate')->where('cid', $id)->count();
+        if(!empty($count)){
+            return response()->json([
+                'msg' => -1
+            ]);
+        }
         DB::table('cfg_coop_cate')->where('id', $id)->delete();
         return response()->json([
             'msg' => 1
